requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
authlib==1.3.0
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
itsdangerous==2.1.2
mako==1.3.10
markupsafe==3.0.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-settings==2.1.0
pyjwt==2.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyyaml==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.23
starlette==0.27.0
typing-extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
